housekeeping: reorganize distribution analysis forms
