+++ src/view/widgets/capture_line_segment.py
@@ -1,8 +1,4 @@
-"""
-"""
-from PyQt5.QtCore import Qt
-
-from qgis.PyQt.QtCore import pyqtSignal
+from qgis.PyQt.QtCore import Qt, pyqtSignal
 from qgis.core import Qgis, QgsWkbTypes, QgsPoint, QgsLineString
 from qgis.gui import QgsMapToolEmitPoint, QgsMapTool, QgsRubberBand
